+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOSharedMailbox/MSFT_EXOSharedMailbox.psm1
@@ -25,6 +25,10 @@ function Get-TargetResource
         $EmailAddresses,
 
         [Parameter()]
+        [System.Boolean]
+        $AuditEnabled,
+
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
         $Ensure = 'Present',
@@ -93,6 +97,7 @@ function Get-TargetResource
                         -RecipientTypeDetails 'SharedMailbox' `
                         -ResultSize Unlimited `
                         -ErrorAction Stop
+                        -PropertySets All
                 }
 
                 if ($null -eq $mailbox)
@@ -101,6 +106,7 @@ function Get-TargetResource
                         -RecipientTypeDetails 'SharedMailbox' `
                         -ResultSize Unlimited `
                         -ErrorAction Stop
+                        -PropertySets All
                 }
             }
             catch
@@ -137,6 +143,7 @@ function Get-TargetResource
             Identity              = $mailbox.Identity
             PrimarySMTPAddress    = $mailbox.PrimarySMTPAddress.ToString()
             Alias                 = $mailbox.Alias
+            AuditEnabled          = $mailbox.AuditEnabled
             EmailAddresses        = $CurrentEmailAddresses
             Ensure                = 'Present'
             Credential            = $Credential
@@ -188,6 +195,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String[]]
         $EmailAddresses = @(),
+
+        [Parameter()]
+        [System.Boolean]
+        $AuditEnabled,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -373,6 +384,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String[]]
         $EmailAddresses,
+
+        [Parameter()]
+        [System.Boolean]
+        $AuditEnabled,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]